requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
django-filter>=24.3,<25.0
Pillow>=10.0,<12.0
